Remove deferred function logic from fileStores
